Rename component/props.
