+++ src/styles/components/data-list.scss
@@ -14,3 +14,7 @@
     overflow-y: auto;
   }
 }
+
+.scroll-container {
+  overflow-y: auto;
+}
